refactor(data): Rename manager to data_manager
