+++ pisek/jobs/task_pipeline.py
@@ -32,7 +32,7 @@ from pisek.task_jobs.task_manager import (
 
 from pisek.jobs.jobs import JobManager
 from pisek.task_jobs.tools import ToolsManager
-from pisek.task_jobs.data.manager import DataManager
+from pisek.task_jobs.data.data_manager import DataManager
 from pisek.task_jobs.generator.generator_manager import (
     PrepareGenerator,
     RunGenerator,
